+++ compiler-tests/src/test/data/diagnostic/inject/member/MissingMemberInjectionShouldFailBinding2.kt
@@ -12,8 +12,8 @@ class Implementation() : Base
 abstract class AppSubscope
 
 @ContributesTo(AppSubscope::class)
+@BindingContainer
 interface SubscopeMultibindingModule {
-
   @Binds
   @IntoMap
   @ClassKey(ClassWithoutMembersInjector::class)